KARAF-7775: Move config shell commands in a dedicated bundle
(cherry picked from commit 43fcb6175483864dd7e5e5d4be0e4f9e2963424b)
