+++ config/command/src/main/java/org/apache/karaf/config/command/ConfigCommandSupport.java
