+++ config/command/src/main/java/org/apache/karaf/config/command/InstallCommand.java
