+++ config/command/src/main/java/org/apache/karaf/config/command/ListCommand.java
