+++ config/command/src/main/resources/OSGI-INF/bundle.info
