+++ config/command/src/main/resources/command/edit.txt
