+++ config/command/src/test/java/org/apache/karaf/config/command/UpdateCommandTest.java
