+++ config/core/src/main/java/org/apache/karaf/config/core/ConfigMBean.java
