+++ config/core/src/main/java/org/apache/karaf/config/core/impl/osgi/Activator.java
