+++ config/core/src/main/resources/OSGI-INF/metatype/metatype.xml
