+++ config/core/src/test/java/org/apache/karaf/config/core/impl/KarafConfigurationPluginTest.java
