+++ config/core/src/test/resources/org.example.Arrays~no_update_required.cfg.json
