+++ config/core/src/test/resources/org.example.Arrays~update_required.cfg.json
