+++ config/core/src/test/resources/test.cfg
